python-peer/requirements.txt
py-libp2p>=0.1.5
aiohttp>=3.8.4
async-timeout>=4.0.2
protobuf>=4.22.0
multiaddr>=0.0.9
click>=8.1.3
rich>=13.3.5
aioconsole>=0.6.0
cryptography>=40.0.0